Updating to libindy 1.6.4 and libsovtoken 0.9.2

+++ vcx/ci/scripts/androidPackage.sh
@@ -10,7 +10,9 @@ do
         arch_folder="armeabi-v7a"
     fi
     mkdir -p ${ANDROID_JNI_LIB}/${arch_folder}
-    cp -v runtime_android_build/libvcx_${arch}/libvcx.so ${ANDROID_JNI_LIB}/${arch_folder}/libvcx.so
+    if [ -e runtime_android_build/libvcx_${arch}/libvcx.so ]; then
+        cp -v runtime_android_build/libvcx_${arch}/libvcx.so ${ANDROID_JNI_LIB}/${arch_folder}/libvcx.so
+    fi
 done
 
 pushd vcx/wrappers/java/android
